adde delete, carousel view to posts

+++ boards/views.py
@@ -200,14 +200,14 @@ class PostUpdateView(UpdateView):
 #         post.updated_at = timezone.now()
 #         post.save()
 #         return redirect('boards:topic_posts', board_pk=post.topic.board.pk, topic_pk=post.topic.pk)
-@login_required
-def delete_post(request, board_pk, topic_pk, post_pk):    
+@login_required()
+def delete_post(request, topic_pk, post_pk):
     post = get_object_or_404(
         Post, 
-        topic__pk=topic_pk,
-        created_by=request.user,
-        pk=post_pk
+        topic__pk = topic_pk,
+        pk = post_pk
     )
-    post.delete()
+    if post.created_by == request.user:
+        post.delete()
 
     return redirect('boards:topic_posts', board_pk=post.topic.board.pk, topic_pk=post.topic.pk)
